move api layer to seperate module can be used to have one library for all series
