+++ dvbapi.h
@@ -1,0 +1,30 @@
+/**
+ * Copyright (c) 2016 harlequin
+ * https://github.com/harlequin/samygo-plugin-dvbapi
+ *
+ * This file is part of samygo-plugin-dvbapi.
+ *
+ * samygo-plugin-dvbapi is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef DVBAPI_H_
+#define DVBAPI_H_
+
+#include "types.h"
+
+void socket_send_pm_table(pmt_t *pmt);
+void socket_send_filter_data(u8 demux_id, u8 filter_num, u8 *data, u32 len);
+
+
+#endif /* DVBAPI_H_ */
